Moved Diff even earlier
