+++ src/strings/diff.scala
@@ -14,9 +14,7 @@
    ║ See the License for the specific language governing permissions and limitations under the License.        ║
    ╚═══════════════════════════════════════════════════════════════════════════════════════════════════════════╝
 */
-package fury.core
-
-import fury.strings._
+package fury.strings
 
 import magnolia._
